+++ src/NLightning.Common/AssemblyInfo.cs
@@ -5,4 +5,5 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("NLightning.Blazor.Tests")]
 [assembly: InternalsVisibleTo("NLightning.BlazorTestApp")]
 [assembly: InternalsVisibleTo("NLightning.Bolts")]
+[assembly: InternalsVisibleTo("NLightning.NLTG")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]