+++ src/NLightning.Common/Crypto/Providers/Native/Constants/ChaCha20Constants.cs
@@ -1,0 +1,9 @@
+namespace NLightning.Common.Crypto.Providers.Native.Constants;
+
+public static class XChaCha20Constants
+{
+    public const int KEY_SIZE = 32;
+    public const int NONCE_SIZE = 24;
+    public const int STATE_SIZE = 16;
+    public const int SUBKEY_SIZE = 32;
+}