+++ src/NLightning.Common/Interfaces/ISecureKeyManager.cs
@@ -1,0 +1,10 @@
+using NBitcoin;
+
+namespace NLightning.Common.Interfaces;
+
+public interface ISecureKeyManager
+{
+    ExtKey GetNextKey(out uint index);
+    Key GetNodeKey();
+    PubKey GetNodePubKey();
+}